// frontend/app/src/models/chart.tsx
// Generated by https://quicktype.io

import { UTCTimestamp } from "lightweight-charts";

export interface IChart {
    time:  UTCTimestamp;
    open:  number;
    high:  number;
    low:   number;
    close: number;
    errors?: ChartError
}

export interface IRequestChart {
    fromToken: string,
    toToken: string,
    interval: number,
    since?: string,
    till?: string
}

export interface ChartError {
    errors: Error[];
}

export interface Error {
    extensions: Extensions;
    locations:  Location[];
    message:    string;
}

export interface Extensions {
    problems: Problem[];
    value:    string;
}

export interface Problem {
    explanation: string;
    path:        any[];
}

export interface Location {
    column: number;
    line:   number;
}