// frontend/app/src/models/trade.tsx
// Generated by https://quicktype.io

export interface ITrade {
    data: Data;
}

export interface Data {
    ethereum: Ethereum;
}

export interface Ethereum {
    dexTrades: DexTrade[];
}

export interface DexTrade {
    timeInterval:  TimeInterval;
    baseCurrency:  ECurrency;
    quoteCurrency: ECurrency;
    tradeAmount:   number;
    trades:        number;
    quotePrice:    number;
    maximum_price: number;
    minimum_price: number;
    open_price:    string;
    close_price:   string;
}

export interface ECurrency {
    symbol:  Symbol;
    address: Address;
}

export enum Address {
    The0Xbb4Cdb9Cbd36B01Bd1Cbaebf2De08D9173Bc095C = "0xbb4cdb9cbd36b01bd1cbaebf2de08d9173bc095c",
    The0Xe9E7Cea3Dedca5984780Bafc599Bd69Add087D56 = "0xe9e7cea3dedca5984780bafc599bd69add087d56",
}

export enum Symbol {
    Busd = "BUSD",
    Wbnb = "WBNB",
}

export interface TimeInterval {
    minute: string;
}